+++ src/main/java/io/github/thebusybiscuit/slimefun4/implementation/tasks/JetBootsTask.java
@@ -4,40 +4,45 @@ import java.util.concurrent.ThreadLocalRandom;
 
 import org.bukkit.Bukkit;
 import org.bukkit.Effect;
+import org.bukkit.Material;
 import org.bukkit.Sound;
 import org.bukkit.entity.Player;
 import org.bukkit.util.Vector;
 
 import io.github.thebusybiscuit.cscorelib2.math.DoubleHandler;
-import me.mrCookieSlime.Slimefun.api.energy.ItemEnergy;
+import io.github.thebusybiscuit.slimefun4.implementation.items.electric.gadgets.JetBoots;
 
 public class JetBootsTask extends PlayerTask {
 
-    private final double speed;
+    private static final float COST = 0.075F;
 
-    public JetBootsTask(Player p, double speed) {
+    private final JetBoots boots;
+
+    public JetBootsTask(Player p, JetBoots boots) {
         super(p);
-        this.speed = speed;
+        this.boots = boots;
     }
 
     @Override
     protected void executeTask() {
-        float cost = 0.075F;
-        float charge = ItemEnergy.getStoredEnergy(p.getInventory().getBoots());
-        double accuracy = DoubleHandler.fixDouble(speed - 0.7);
+        if (p.getInventory().getBoots() == null || p.getInventory().getBoots().getType() == Material.AIR) {
+            return;
+        }
 
-        if (charge >= cost) {
-            p.getInventory().setBoots(ItemEnergy.chargeItem(p.getInventory().getBoots(), -cost));
+        double accuracy = DoubleHandler.fixDouble(boots.getSpeed() - 0.7);
 
+        if (boots.removeItemCharge(p.getInventory().getBoots(), COST)) {
             p.getWorld().playSound(p.getLocation(), Sound.ENTITY_TNT_PRIMED, (float) 0.25, 1);
             p.getWorld().playEffect(p.getLocation(), Effect.SMOKE, 1, 1);
             p.setFallDistance(0F);
             double gravity = 0.04;
             double offset = ThreadLocalRandom.current().nextBoolean() ? accuracy : -accuracy;
-            Vector vector = new Vector(p.getEyeLocation().getDirection().getX() * speed + offset, gravity, p.getEyeLocation().getDirection().getZ() * speed - offset);
+            Vector vector = new Vector(p.getEyeLocation().getDirection().getX() * boots.getSpeed() + offset, gravity, p.getEyeLocation().getDirection().getZ() * boots.getSpeed() - offset);
 
             p.setVelocity(vector);
         }
-        else Bukkit.getScheduler().cancelTask(id);
+        else {
+            Bukkit.getScheduler().cancelTask(id);
+        }
     }
 }